removed old card component and replace with optional clickable card
