+++ components/sections/stats/NetPosition.jsx
@@ -1,15 +1,11 @@
 import Card from "/components/card/Card";
-import { useEffect, useState } from "react";
-import { getUserStats } from "/utils/data/users";
 import Money from "/components/text/money/Money";
-import TextWithButton from "/components/text/title/TextWithButton";
 import TextWithTitle from "components/text/title/TextWithTitle";
 import { useSession } from "next-auth/react";
 import styles from "./NetPosition.module.scss";
 import { NetPositionPlaceholder } from "/components/placeholders/Placeholders";
 import { useUserStats } from "/utils/hooks";
 import Spinner from "/components/placeholders/spinner/Spinner";
-import ClickableCard from "/components/card/ClickableCard";
 
 export default function NetPosition() {
 	const { data: session } = useSession();
@@ -29,7 +25,7 @@ export default function NetPosition() {
 				<>
 					{stats.current.net != 0 && (
 						<div className={styles.wrapper}>
-							<ClickableCard
+							<Card
 								className={styles.netWrapper}
 								action={
 									<p className={styles.action}>
@@ -42,9 +38,9 @@ export default function NetPosition() {
 									notColoured
 									includeSign
 								/>
-							</ClickableCard>
+							</Card>
 							<div className={styles.individualDebtTotals}>
-								<ClickableCard
+								<Card
 									className={styles.debtWrapper}
 									href="/debts"
 									action={
@@ -55,9 +51,9 @@ export default function NetPosition() {
 									reverseAction
 									includeArrow>
 									<Money amount={stats.current.debt} />
-								</ClickableCard>
+								</Card>
 
-								<ClickableCard
+								<Card
 									className={styles.unreceivedWrapper}
 									href="/unreceived-payments"
 									action={
@@ -68,7 +64,7 @@ export default function NetPosition() {
 									reverseAction
 									includeArrow>
 									<Money amount={stats.current.unreceived} />
-								</ClickableCard>
+								</Card>
 							</div>
 						</div>
 					)}
